Unbound: G_FOGF float fog factor
The G_MW_FOG moveword packs the fog multiplier/offset as s16, which is
why OoT clamps fogNear to 996 and fog can never start past ~2500 units.
G_FOGF (0x41) carries the same two numbers as floats (w1 -> {mul, off});
fog_mul/fog_offset in the RSP state are floats. Vanilla G_MW_FOG is
unchanged.

+++ include/fast/interpreter.h
@@ -242,7 +242,7 @@ struct RSP {
     bool lights_changed;
 
     uint32_t geometry_mode;
-    int16_t fog_mul, fog_offset;
+    float fog_mul, fog_offset; // SOH [Unbound] float so fog can start beyond the s16 range (OTR_G_FOGF)
 
     uint32_t extra_geometry_mode;
